64-bit compile fix.

+++ source/ddox/htmlgenerator.d
@@ -29,7 +29,7 @@ void generateHtmlDocs(Path dst_path, Package root)
 {
 	auto settings = new GeneratorSettings;
 
-	string linkTo(Entity ent, int level)
+	string linkTo(Entity ent, size_t level)
 	{
 		auto dst = appender!string();
 		if( level ) foreach( i; 0 .. level ) dst.put("../");
